standardize sensor sources, provide a method to switch between map and odom
