+++ src/vtk/DRCFilters/vtkRosDepthImageSubscriber.cpp
@@ -12,7 +12,7 @@ vtkRosDepthImageSubscriber::vtkRosDepthImageSubscriber()
   }
   tf_listener_ = boost::make_shared<tf::TransformListener>();
   dataset_ = vtkSmartPointer<vtkPolyData>::New();
-  fixed_frame_ = "map"; // or "map"
+  fixed_frame_ = "map"; // or "odom"
   sec_ = 0;
   nsec_ = 0;
 }
