+++ src/vtk/DRCFilters/vtkRosGridMapSubscriber.cpp
@@ -1,4 +1,5 @@
 #include "vtkRosGridMapSubscriber.h"
+#include <transformPolyDataUtils.h>
 
 #include "vtkTransform.h"
 #include "vtkTransformPolyDataFilter.h"
@@ -23,6 +24,8 @@ vtkRosGridMapSubscriber::vtkRosGridMapSubscriber()
   if (!ros::isInitialized()) {
     std::cout << "WARNING: vtkRosGridMapSubscriber: ROS not Initialized\n";
   }
+  tf_listener_ = boost::make_shared<tf::TransformListener>();
+  fixed_frame_ = "map"; // or "odom"
 }
 
 vtkRosGridMapSubscriber::~vtkRosGridMapSubscriber() {
@@ -50,9 +53,24 @@ void vtkRosGridMapSubscriber::GridMapCallback(const grid_map_msgs::GridMap& mess
   // Convert message to map.
   grid_map::GridMapRosConverter::fromMessage(message, inputMap_);
 
+  std::string frame_id = message.info.header.frame_id;
+  vtkSmartPointer<vtkTransform> sensorToLocalTransform = vtkSmartPointer<vtkTransform>::New();
+  tf::StampedTransform transform;
+  ros::Time time = message.info.header.stamp;
+  tf_listener_->waitForTransform(fixed_frame_, frame_id, time, ros::Duration(2.0));
+  try {
+    tf_listener_->lookupTransform(fixed_frame_, frame_id, time, transform);
+    sensorToLocalTransform = transformPolyDataUtils::transformFromPose(transform);
+  }
+  catch (tf::TransformException& ex){
+    ROS_ERROR("%s",ex.what());
+    return;
+  }
+
   //we can't modify dataset_ if it's being copied in GetMesh
   std::lock_guard<std::mutex> lock(mutex_);
   dataset_ = ConvertMesh();
+  transformPolyDataUtils::transformPolyData(dataset_, dataset_, sensorToLocalTransform);  
 }
 
 vtkSmartPointer<vtkPolyData> vtkRosGridMapSubscriber::ConvertMesh()
